+++ operations/cache.go
@@ -64,8 +64,9 @@ func cacheLocation() string {
 
 func SummonCache() (*Cache, error) {
 	var result Cache
-	completed := pathlib.LockWaitMessage("Serialized cache access [cache lock]")
-	locker, err := pathlib.Locker(cacheLockFile(), 125)
+	lockfile := cacheLockFile()
+	completed := pathlib.LockWaitMessage(lockfile, "Serialized cache access [cache lock]")
+	locker, err := pathlib.Locker(lockfile, 125)
 	completed()
 	if err != nil {
 		return nil, err
@@ -86,8 +87,9 @@ func SummonCache() (*Cache, error) {
 }
 
 func (it *Cache) Save() error {
-	completed := pathlib.LockWaitMessage("Serialized cache access [cache lock]")
-	locker, err := pathlib.Locker(cacheLockFile(), 125)
+	lockfile := cacheLockFile()
+	completed := pathlib.LockWaitMessage(lockfile, "Serialized cache access [cache lock]")
+	locker, err := pathlib.Locker(lockfile, 125)
 	completed()
 	if err != nil {
 		return err