+++ xviper/wrapper.go
@@ -38,7 +38,7 @@ func (it *config) Save() {
 	if len(it.Filename) == 0 {
 		return
 	}
-	completed := pathlib.LockWaitMessage("Serialized config access [config lock]")
+	completed := pathlib.LockWaitMessage(it.Lockfile, "Serialized config access [config lock]")
 	locker, err := pathlib.Locker(it.Lockfile, 125)
 	completed()
 	if err != nil {
@@ -59,7 +59,7 @@ func (it *config) Save() {
 }
 
 func (it *config) Reload() {
-	completed := pathlib.LockWaitMessage("Serialized config access [config lock]")
+	completed := pathlib.LockWaitMessage(it.Lockfile, "Serialized config access [config lock]")
 	locker, err := pathlib.Locker(it.Lockfile, 125)
 	completed()
 	if err != nil {